// src/common/types/absenteeism.types.ts
export interface AusentismoData {
  FTC: number[];
  TNL: number[];
  FTMU: number[];
  FTU: number[];
  FTNU: number[];
  Enfermedad: number[];
  AsuntosPropios: number[];
  AccidenteTrabajo: number[];
  AccidenteEquiparado: number[];
  AusenciasInjustificadas: number[];
  PromedioTabla: any[];
  PromedioAnterior: number[];
}